Define ProjectsEditorWindow to enable configuring Projects for a Roadmap
Because RoadmapConfiguratorViewModel doesn't actually expose a collection of projects yet, the configured set of projects are flattened into a single number of work items.
User experience of ProjectsEditorWindow's DataGrid is suboptimal, as value validation doesn't give a descriptive message to the user yet.

+++ Kanban Project Management App/Domain/Project.cs
@@ -35,7 +35,7 @@ namespace KanbanProjectManagementApp.Domain
 
         public bool HasWorkToBeCompleted => NumberOfWorkItemsRemaining > 0;
 
-        public double NumberOfWorkItemsRemaining { get; private set; }
+        public int NumberOfWorkItemsRemaining { get; private set; }
 
         /// <exception cref="InvalidOperationException">Thrown when <see cref="HasWorkToBeCompleted"/> is false.</exception>
         public void CompleteWorkItem()
